fix(ent/schema): use auto incrementing id when storing timeline events
fixes #5.

we need to be able to replicate timelines in the order that they
happened; by using an auto-incrementing ID, we can ensure that the order
in the `timeline_events` table is the same as the order in which the
events were inserted.

this also allows us to correctly store events like cross-referenced,
which have neither a id nor a node_id.